Updated volunteers and researchers sign in pages
Users who already signed in as a volunteer, are not allowrd to sign in as a researcher. Users who already signed in as a researcher, are not allowed to sign in as a researcher again nor as a volunteer. This way, each user will sign in as a unique identity.