Updated User Search Page and Added Some MySQL Functions on Boincuser class
Previously, the user search page does searching based on the user screen name. Now, it will run based on user's anonymized name. Besides that, I used BOINC mechanism to communicate with the database and add some functions to class Boincuser so that the object of that class could give an access to screen_name_anonymization table.